Fixed cv2 not liking the type of the camera image

+++ src/surface/gui/gui/widgets/video_widget.py
@@ -162,7 +162,7 @@ class VideoWidget(QWidget):
         """Convert from an opencv image to QPixmap."""
         if self.camera_description.type == CameraType.ETHERNET:
             # Switches ethernet's color profile from BayerBGR to BGR
-            cv_img = cv2.cvtColor(cv_img.astype(int), cv2.COLOR_BAYER_BGGR2BGR)
+            cv_img = cv2.cvtColor(cv_img.astype("uint8"), cv2.COLOR_BAYER_BGGR2BGR)
 
         # Color image
         if len(cv_img.shape) == COLOR:
